refactor: centralize file processing helpers
