+++ services/data_processing/__init__.py
@@ -7,6 +7,7 @@ from .file_handler import FileHandler, process_file_simple
 # the new class here and optionally expose it under the old name for backward
 # compatibility.
 from .file_processor import UnicodeFileProcessor
+from .base_file_processor import BaseFileProcessor
 
 # Provide the legacy name ``FileProcessor`` for callers that still import it
 # from ``services.data_processing``.
@@ -73,4 +74,5 @@ __all__ = [
     "FileFormatError",
     "FileSizeError",
     "FileSecurityError",
+    "BaseFileProcessor",
 ]
